backend/pyproject.toml
[project]
name = "ohif-backend"
version = "0.1.0"
description = "OHIF Cornerstone Viewer Backend - Django + DRF"
requires-python = ">=3.12"
readme = "README.md"

dependencies = [
    "asgiref==3.8.1",
    "Django==5.0.4",
    "djangorestframework==3.15.1",
    "djangorestframework-simplejwt==5.3.1",
    "django-cors-headers==4.3.1",
    "django-environ==0.11.2",
    "django-jazzmin",
    "psycopg[binary]==3.2.4",
    "celery==5.3.6",
    "redis==5.0.4",
    "django-redis==5.4.0",
    "httpx==0.27.0",
    "Pillow==11.2.1",
    "gunicorn==22.0.0",
    "whitenoise==6.6.0",
    "sqlparse==0.5.3",
    "python-decouple==3.8",
    "uvicorn[standard]",
]

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]

[tool.uv]
managed = true

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-django",
    "ruff",
    "black",
]

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"